fastapi_agent/requirements.txt
fastapi
uvicorn[standard]
PyGObject
requests
netifaces
kafka-python